api/requirements.txt
flask-restful==0.3.6
pyyaml
certifi
ipython
pika==0.13.1
simplejson==3.16.0
flask==1.0.2
ipaddress
six
urllib3
mongoengine==0.17.0
pymysql
google-auth
adal
setuptools
python-dateutil
kubernetes==9.0.0
requests-oauthlib
websocket-client
pymongo==3.7.2
raven==6.9.0

Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pymongo==3.7.2
redis==3.2.1
Werkzeug==0.15.2
pyotp===2.3.0